create comment component
