+++ client/src/components/CreateComment.js
@@ -2,47 +2,7 @@ import React, { Component } from 'react'
 
 const url = 'http://localhost:1340/comment';
 
-export default class CreateComment extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            content: ""
-        }
-    }
-
-    onFormChange = event => {
-        const { name, value } = event.target;
-        this.setState({
-            [name]: value
-        })
-    }
-
-    onFormSubmit = event => {
-        event.preventDefault();
-        let data = {
-            content: this.state.content
-
-        }
-        fetch(url, {
-            method: 'POST',
-            body: JSON.stringify(data),
-            headers: {
-                'Content-Type': 'application/json'
-            }
-        }).then(response => {
-            return response.json();
-        })
-        this.setState({
-            title: "",
-            content: ""
-        })
-
-    }
-
-
-
-
-
+export default class CreateComment extends Component {  
 
 
     render() {
@@ -50,8 +10,8 @@ export default class CreateComment extends Component {
             <div className='createCommentPage'>
                 <h2 className='title'>Comment</h2>
                 <h3>Respond to this post</h3>
-                <form onSubmit={this.onFormSubmit} id="thread-form">
-                    <textarea rows='4' cols='40' className='commentbox' value={this.state.content} onChange={this.onFormChange} name='content' placeholder='Enter comments here'></textarea>
+                <form onSubmit={this.props.onFormSubmit} id="thread-form">
+                    <textarea rows='4' cols='40' className='commentbox' value={this.props.comment} onChange={this.props.onFormChange} name='content' placeholder='Enter comments here'></textarea>
                     <button>Comment</button>
 
                 </form>
